+++ src/main/views/viewManager.test.js
@@ -6,6 +6,8 @@
 
 import {dialog} from 'electron';
 
+import ServerViewState from 'app/serverViewState';
+
 import {BROWSER_HISTORY_PUSH, LOAD_SUCCESS, SET_ACTIVE_VIEW} from 'common/communication';
 import {TAB_MESSAGING} from 'common/views/View';
 import ServerManager from 'common/servers/serverManager';
@@ -29,6 +31,11 @@ jest.mock('electron', () => ({
         on: jest.fn(),
         handle: jest.fn(),
     },
+}));
+jest.mock('app/serverViewState', () => ({
+    getCurrentServer: jest.fn(),
+    updateCurrentView: jest.fn(),
+    init: jest.fn(),
 }));
 jest.mock('common/views/View', () => ({
     getViewName: jest.fn((a, b) => `${a}-${b}`),
@@ -70,13 +77,11 @@ jest.mock('main/windows/mainWindow', () => ({
     on: jest.fn(),
 }));
 jest.mock('common/servers/serverManager', () => ({
-    getCurrentServer: jest.fn(),
     getOrderedTabsForServer: jest.fn(),
     getAllServers: jest.fn(),
     hasServers: jest.fn(),
     getLastActiveServer: jest.fn(),
     getLastActiveTabForServer: jest.fn(),
-    updateLastActive: jest.fn(),
     lookupViewByURL: jest.fn(),
     getRemoteInfo: jest.fn(),
     on: jest.fn(),
@@ -351,7 +356,7 @@ describe('main/views/viewManager', () => {
             viewManager.showById = jest.fn();
             MainWindow.get.mockReturnValue(window);
             ServerManager.hasServers.mockReturnValue(true);
-            ServerManager.getCurrentServer.mockReturnValue({id: 'server-0'});
+            ServerViewState.getCurrentServer.mockReturnValue({id: 'server-0'});
         });
 
         afterEach(() => {
@@ -440,7 +445,7 @@ describe('main/views/viewManager', () => {
 
         beforeEach(() => {
             ServerManager.getAllServers.mockReturnValue(servers);
-            ServerManager.getCurrentServer.mockReturnValue(servers[0]);
+            ServerViewState.getCurrentServer.mockReturnValue(servers[0]);
             urlUtils.cleanPathName.mockImplementation((base, path) => path);
         });
